+++ addOns/authhelper/src/main/java/org/zaproxy/addon/authhelper/AuthTestDialog.java
@@ -43,6 +43,7 @@ import org.parosproxy.paros.model.Session;
 import org.parosproxy.paros.view.View;
 import org.zaproxy.addon.authhelper.AutoDetectSessionManagementMethodType.AutoDetectSessionManagementMethod;
 import org.zaproxy.addon.authhelper.BrowserBasedAuthenticationMethodType.BrowserBasedAuthenticationMethod;
+import org.zaproxy.addon.authhelper.internal.StepsPanel;
 import org.zaproxy.addon.pscan.ExtensionPassiveScan2;
 import org.zaproxy.zap.ZAP;
 import org.zaproxy.zap.authentication.AuthenticationMethod;
@@ -103,6 +104,8 @@ public class AuthTestDialog extends StandardFieldsDialog {
     private JLabel sessionIdLabel = new JLabel();
     private JLabel verifIdLabel = new JLabel();
 
+    private StepsPanel stepsPanel;
+
     private ZapTextArea diagnosticField;
     private Boolean usernameFieldFound;
     private Boolean passwordFieldFound;
@@ -115,7 +118,9 @@ public class AuthTestDialog extends StandardFieldsDialog {
                 "authhelper.auth.test.dialog.title",
                 DisplayUtils.getScaledDimension(600, 480),
                 new String[] {
-                    "authhelper.auth.test.dialog.tab.test", "authhelper.auth.test.dialog.tab.diag"
+                    "authhelper.auth.test.dialog.tab.test",
+                    "authhelper.auth.test.dialog.tab.steps",
+                    "authhelper.auth.test.dialog.tab.diag"
                 });
 
         this.ext = ext;
@@ -140,7 +145,13 @@ public class AuthTestDialog extends StandardFieldsDialog {
         this.addCustomComponent(0, getResultsPanel());
         this.addPadding(0);
 
-        this.addMultilineField(1, DIAGNOSTICS_LABEL, "");
+        int tab = 1;
+        stepsPanel = new StepsPanel(this, true);
+        stepsPanel.setSteps(params.getSteps());
+        setCustomTabPanel(tab, stepsPanel.getPanel());
+
+        tab++;
+        addMultilineField(tab, DIAGNOSTICS_LABEL, "");
         diagnosticField = (ZapTextArea) this.getField(DIAGNOSTICS_LABEL);
         diagnosticField.setEditable(false);
         ext.setAuthDiagCollectorOutput(diagnosticField);
@@ -160,7 +171,7 @@ public class AuthTestDialog extends StandardFieldsDialog {
         buttonPanel.add(copyButton, LayoutHelper.getGBC(1, 0, 1, 0.3D));
         buttonPanel.add(new JLabel(), LayoutHelper.getGBC(2, 0, 1, 0.3D));
 
-        this.addCustomComponent(1, COPY_LABEL, buttonPanel);
+        addCustomComponent(tab, COPY_LABEL, buttonPanel);
 
         ZapTextField text = (ZapTextField) this.getField(LOGIN_URL_LABEL);
         text.setText(params.getLoginUrl());
@@ -263,6 +274,7 @@ public class AuthTestDialog extends StandardFieldsDialog {
             String browserId = ((BrowserUI) browserCombo.getSelectedItem()).getBrowser().getId();
             am.setBrowserId(browserId);
             am.setLoginPageWait(this.getIntValue(WAIT_LABEL));
+            am.setAuthenticationSteps(stepsPanel.getSteps());
             reloadAuthenticationMethod(am);
             context.setAuthenticationMethod(am);
 
@@ -451,6 +463,7 @@ public class AuthTestDialog extends StandardFieldsDialog {
         params.setBrowser(((BrowserUI) browserCombo.getSelectedItem()).getBrowser().getId());
         params.setWait(this.getIntValue(WAIT_LABEL));
         params.setDemoMode(this.getBoolValue(DEMO_LABEL));
+        params.setSteps(stepsPanel.getSteps());
     }
 
     @Override