+++ addOns/authhelper/src/main/java/org/zaproxy/addon/authhelper/AuthhelperParam.java
@@ -19,7 +19,10 @@
  */
 package org.zaproxy.addon.authhelper;
 
+import java.util.List;
+import java.util.Objects;
 import org.parosproxy.paros.common.AbstractParam;
+import org.zaproxy.addon.authhelper.internal.AuthenticationStep;
 import org.zaproxy.zap.extension.selenium.Browser;
 
 public class AuthhelperParam extends AbstractParam {
@@ -31,12 +34,14 @@ public class AuthhelperParam extends AbstractParam {
     private static final String BROWSER_KEY = AUTO_KEY + ".browser";
     private static final String WAIT_KEY = AUTO_KEY + ".wait";
     private static final String DEMO_MODE_KEY = AUTO_KEY + ".demo";
+    private static final String STEP_KEY = AUTO_KEY + ".steps.step";
 
     private String loginUrl;
     private String username;
     private String browser;
     private int wait = 2;
     private boolean demoMode;
+    private List<AuthenticationStep> steps = List.of();
 
     public AuthhelperParam() {}
 
@@ -47,6 +52,13 @@ public class AuthhelperParam extends AbstractParam {
         this.browser = this.getString(BROWSER_KEY, Browser.FIREFOX.getId());
         this.wait = getInteger(WAIT_KEY, 2);
         this.demoMode = getBoolean(DEMO_MODE_KEY, false);
+
+        steps =
+                getConfig().getList(STEP_KEY).stream()
+                        .map(Object::toString)
+                        .map(AuthenticationStep::decode)
+                        .filter(Objects::nonNull)
+                        .toList();
     }
 
     public String getLoginUrl() {
@@ -92,5 +104,18 @@ public class AuthhelperParam extends AbstractParam {
     public void setWait(int wait) {
         this.wait = wait;
         getConfig().setProperty(WAIT_KEY, wait);
+    }
+
+    public List<AuthenticationStep> getSteps() {
+        return steps;
+    }
+
+    public void setSteps(List<AuthenticationStep> steps) {
+        this.steps = steps == null ? List.of() : steps;
+
+        getConfig().clearProperty(STEP_KEY);
+        this.steps.stream()
+                .map(AuthenticationStep::encode)
+                .forEach(e -> getConfig().addProperty(STEP_KEY, e));
     }
 }