+++ addOns/authhelper/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - Added support for Client Script Authentication when used in conjunction with the Ajax Spider add-on.
+- Add support for custom authentication steps in Browser Based Authentication.
 
 ## [0.18.0] - 2025-01-27
 ### Changed