+++ addOns/automation/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Added support for Client Script Authentication when the Ajax Spider is used in conjunction with the Auth Helper add-on.
+- Add support for custom authentication steps in Browser Based Authentication.
 
 ## [0.44.0] - 2025-01-09
 ### Added